% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_parms.R
\name{edit_parms}
\alias{edit_parms}
\title{Edit an Agro-IBIS parameter file}
\usage{
edit_parms(crop.parms.lst,
  table.name = c("crop_growth_physiology_properties",
  "crop_growth_leafarea_c_allocation", "climatic_manage_planting_control",
  "gdd_phenology_control", "wheat_growth_control", "misc_crop_control",
  "sugarcane_control", "crop_residue_control"), col = NULL, row = NULL,
  value = 0)
}
\arguments{
\item{crop.parms.lst}{list wth components as produced by 'read_crop_parms'}

\item{table.name}{name of the table to edit}

\item{col}{name (or index) of column to edit}

\item{row}{name (or index) of row to edit}

\item{value}{value for replacement}
}
\value{
list structure with replaced values
}
\description{
Edit an Agro-IBIS parameter file
}
\examples{
\donttest{
extd.dir <- system.file("extdata", package = "agroibis")
cpp <- read_crop_parms(file = "params_text.crp", src.dir = extd.dir)
cpp[["wheat_growth_control"]]
cpp2 <- edit_parms(cpp, table.name = "wheat_growth_control",
                   col = 2, row = 1, value = 0.812)
cpp2[["wheat_growth_control"]]
}

}
