% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_crop_parms.R
\name{read_crop_parms}
\alias{read_crop_parms}
\title{Read an Agro-IBIS crop parameter file}
\usage{
read_crop_parms(file, src.dir = ".", verbose = TRUE)
}
\arguments{
\item{file}{file name, should end in .crp}

\item{src.dir}{source directory}

\item{verbose}{whether to print progress messages (default = TRUE)}
}
\value{
list structure with matrices for different sections of the parameter file
}
\description{
Read an Agro-IBIS crop parameter file
}
\examples{
\donttest{
extd.dir <- system.file("extdata", package = "agroibis")
cpp <- read_crop_parms(file = "params_text.crp", src.dir = extd.dir)
}

}
