% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_crop_parms.R
\name{write_crop_parms}
\alias{write_crop_parms}
\title{Write an Agro-IBIS crop parameter file}
\usage{
write_crop_parms(crop.parms.lst, file = "params_text.crp", wrt.dir = ".")
}
\arguments{
\item{crop.parms.lst}{list object with the structure are generated by 'read_crop_parms'}

\item{file}{default 'params_text.crp'}

\item{wrt.dir}{write directory}
}
\value{
does not return anything but writes a file to disk
}
\description{
Write an Agro-IBIS crop parameter file
}
\examples{
\donttest{
extd.dir <- system.file("extdata", package = "agroibis")
cpp <- read_crop_parms(file = "params_text.crp", src.dir = extd.dir)
## Edit it using edit_parms
cpp2 <- edit_parms(cpp, table.name = "wheat_growth_control",
                   col = 2, row = 1, value = 0.812)
write_crop_parms(cpp2, file = "params_text.crp")
}

}
